fix: resolve various compilation warnings
1. Fixed unused parameter warnings by removing parameter names in lambda
captures
2. Added Qt version check for deprecated AA_UseHighDpiPixmaps attribute
3. Fixed documentation comment syntax in DFontComboBox
4. Added nullptr check for QGuiApplication::inputMethod()
5. Added parent parameter in DPrintPreviewWidget constructor
6. Fixed pixel metric switch case with static cast
7. Updated meta type registration for Qt6 compatibility
8. Removed unused variable casts in test cases
9. Added Qt version check for QContextMenuEvent constructor

Influence:
1. Verify UI elements still render correctly after DPI attribute changes
2. Test input method behavior on tablet devices
3. Check print preview functionality
4. Validate combo box and text edit test cases still pass
5. Confirm no new warnings appear during compilation

fix: 修复各种编译警告

1. 在lambda捕获中移除未使用的参数名以消除警告
2. 添加Qt版本检查处理废弃的AA_UseHighDpiPixmaps属性
3. 修复DFontComboBox中的文档注释语法
4. 为QGuiApplication::inputMethod()添加nullptr检查
5. 在DPrintPreviewWidget构造函数中添加父控件参数
6. 使用静态转换修复像素度量switch case
7. 更新元类型注册以兼容Qt6
8. 移除测试用例中未使用的变量转换
9. 为QContextMenuEvent构造函数添加Qt版本检查

Influence:
1. 验证DPI属性变更后UI元素仍能正确渲染
2. 测试平板设备上的输入法行为
3. 检查打印预览功能
4. 确保组合框和文本编辑测试用例仍能通过
5. 确认编译时不再出现新的警告

+++ src/widgets/dstyle.cpp
@@ -2194,7 +2194,7 @@ void DStyle::drawControl(QStyle::ControlElement ce, const QStyleOption *opt, QPa
  */
 int DStyle::pixelMetric(QStyle::PixelMetric m, const QStyleOption *opt, const QWidget *widget) const
 {
-    switch (m) {
+    switch (static_cast<uint32_t>(m)) {
     case PM_ButtonDefaultIndicator:
     case PM_ButtonShiftHorizontal:
     case PM_ButtonShiftVertical:
